Add support for new Pi2/3 trailer
According to the official documentation[1], one additional step is required to
get a fully functional kernel. This additional step is a call to the new
`mkknlimg` script (found in the Raspberry Pi kernel source tree), which appends
a so-called 'trailer' to the kernel image. The trailer's purpose is to let the
bootloader know that the kernel is DT-capable so that the bootloader can pass
the appropriate bootloader to it. Without the trailer, the kernel will boot
*without* the DT making it blind to many of the on-board devices.

The pi5_trailer.sh script has been added, which takes the kernel version as its
argument, and runs the correct version of `mkknlimg` script found in the build
directory.

[1] https://www.raspberrypi.org/documentation/linux/kernel/building.md

Signed-off-by: Branko Vukelic <branko@outernet.is>